a2a/a2a-lambda-python-async/src/server/requirements.txt
boto3
langchain-aws>=0.2.35
langgraph-api>= 0.4.9
a2a-sdk
a2a-sdk[http-server]
uvicorn